refactor: decompose main_gpui.rs, fix all lizard violations, remove CI exclusions
Extract system_tray and startup modules from main_gpui.rs (1282 -> 701 lines).
Split main() into focused helpers: run_gpui_app, run_tokio_runtime,
create_services, create_presenter_channels_and_bridges, start_all_presenters.

Fix all 8 lizard function-length violations across src/:
- render_edit_form -> render_label_field + render_value_field + render_form_buttons
- render_conversation_card -> render_card_actions
- create_agent -> do_run_agent_stream
- send_message refactored for line count
- 4 test functions split into smaller focused tests

Remove all --exclude and -not -path flags from CI lizard/file-length checks.
Update coverage ignore regex to account for GPUI render/IME files that
require a live window context and cannot be unit-tested.
Fix NavigationChannel race in api_key_manager delete test.
