+++ .github/workflows/pr-quality-and-e2e.yml
@@ -61,11 +61,7 @@ jobs:
           set -euo pipefail
           . .venv-lizard/bin/activate
 
-          python -m lizard -C 50 -L 100 -w src/ \
-            --exclude "src/main_gpui.rs" \
-            --exclude "src/bin/*" \
-            --exclude "src/services/chat.rs" \
-            --exclude "src/llm/client_agent.rs"
+          python -m lizard -C 50 -L 100 -w src/
 
           error_exit=0
           while IFS= read -r file; do
@@ -76,11 +72,7 @@ jobs:
             elif [ "$lines" -gt 750 ]; then
               echo "WARNING: $file has $lines lines (recommended max 750)"
             fi
-          done < <(
-            find src -name "*.rs" \
-              -not -path "src/main_gpui.rs" \
-              -not -path "src/bin/*" | sort
-          )
+          done < <(find src -name "*.rs" | sort)
 
           if [ "$error_exit" -eq 1 ]; then
             exit 1
